requirements.txt
numpy
face_recognition
mysql-connector-python
opencv-python
setuptools